Fix heads-up display tests - The test that asserts that the profile activity shows up on a button click was failing because this activity uses the DB which was not initialized. Inheriting the test suite from `PolyPolyTest` neatly did the job.
